+++ scripts/scrape_keywords.py
@@ -1,16 +1,17 @@
+from __future__ import annotations
+
 import csv
 from pathlib import Path
 
 import requests
 from bs4 import BeautifulSoup
 
-
 # Databricks's SQL keywords page was last updated October 10, 2023. It is archived at:
 # https://web.archive.org/web/20240122080239/https://docs.databricks.com/en/sql/language-manual/sql-ref-reserved-words.html
 URL = "https://docs.databricks.com/en/sql/language-manual/sql-ref-reserved-words.html"
 
 
-def scrape_keywords() -> None:
+def scrape_keywords() -> list[str]:
     page = requests.get(URL)
     soup = BeautifulSoup(page.content, "html.parser")