+++ src/harlequin_databricks/completions.py
@@ -12,12 +12,12 @@ def load_completions() -> list[HarlequinCompletion]:
     keywords_path = Path(__file__).parent / "keywords.csv"
     with keywords_path.open("r") as file:
         reader = csv.reader(file, dialect="unix")
-        for name in reader:
+        for row in reader:
             completions.append(
                 HarlequinCompletion(
-                    label=name[0].lower(),
+                    label=row[0].lower(),
                     type_label="kw",
-                    value=name[0].lower(),
+                    value=row[0].lower(),
                     priority=1000,
                     context=None,
                 )