+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-02-10
+
+### Features
+
+-   Faster indexing for Unity Catalog assets. Makes loading & refreshing Harlequin's Data Catalog
+pane much quicker for Databricks instances running Unity Catalog. The new method fetches metadata
+for Unity Catalog assets from `system.information_schema`, requiring only two SQL queries to do so.
+This is much faster than the old method, which required separate SQL calls for each table. The old
+method is still in use for indexing legacy metastores (e.g. `hive_metastore`) however, as that
+metadata is not contained in Information schema:
+https://docs.databricks.com/en/sql/language-manual/sql-ref-information-schema.html
+-   Add new command line flag `--skip-legacy-indexing` to skip the indexing of legacy metastores.
+Setting this flag is recommended if your Databricks instance runs Unity Catalog and you do not want
+the overhead of slow indexing of legacy metastores. When this flag is set, only Unity Catalog
+assets will show up in the Data Catalog pane.
+
+### Bug Fixes
+
+-   Fix minor formatting (black), types (mypy) & linting (ruff) issues, and a test failure.
+-   Don't use underscores in CLI option names, i.e. make it so the only acceptable option version
+is the one written with hyphens not underscores.
+
 ## [0.1.1] - 2024-02-08
 
 ### Bug Fixes
@@ -17,7 +39,9 @@ All notable changes to this project will be documented in this file.
 
 -   Adds a Databricks adapter for SQL warehouses and DBR interactive clusters.
 
-[Unreleased]: https://github.com/alexmalins/harlequin-databricks/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/alexmalins/harlequin-databricks/compare/0.2.0...HEAD
+
+[0.2.0]: https://github.com/alexmalins/harlequin-databricks/compare/0.1.1...0.2.0
 
 [0.1.1]: https://github.com/alexmalins/harlequin-databricks/compare/0.1.0...0.1.1